+++ test-utils/src/lib.rs
@@ -1,82 +1,44 @@
 use bitcoin::blockdata::constants::genesis_block;
 use bitcoin::{
-    hashes::Hash,
-    secp256k1,
-    secp256k1::{constants::SECRET_KEY_SIZE, Secp256k1, Signing},
-    util::uint::Uint256,
-    Address, Block as BitcoinBlock, BlockHash, BlockHeader as Header, KeyPair, Network, OutPoint,
-    PublicKey, Script, Transaction, TxIn, TxMerkleNode, TxOut, WScriptHash, Witness,
-    XOnlyPublicKey,
+    hashes::Hash, secp256k1::Secp256k1, util::uint::Uint256, Address, Block as BitcoinBlock,
+    BlockHash, BlockHeader as Header, KeyPair, Network, OutPoint, PublicKey, Script, Transaction,
+    TxIn, TxMerkleNode, TxOut, WScriptHash, Witness, XOnlyPublicKey,
 };
 use ic_btc_types::Block;
-use simple_rng::SimpleRng;
-use std::cell::RefCell;
+use simple_rng::{fill_bytes, generate_keypair};
 use std::str::FromStr;
 
 mod simple_rng;
 
-thread_local! {
-    static RNG: RefCell<simple_rng::SimpleRng> = RefCell::new(simple_rng::SimpleRng::new(std::time::Instant::now().elapsed().as_nanos() as u64));
-}
-
-fn with_rng<F, R>(f: F) -> R
-where
-    F: FnOnce(&mut simple_rng::SimpleRng) -> R,
-{
-    RNG.with(|rng| {
-        let mut rng = rng.borrow_mut();
-        f(&mut rng)
-    })
-}
-
-pub fn generate_keypair<C: Signing>(
-    secp: &Secp256k1<C>,
-    rng: &mut SimpleRng,
-) -> (secp256k1::SecretKey, secp256k1::PublicKey) {
-    let mut data = [0u8; SECRET_KEY_SIZE];
-    rng.fill_bytes(&mut data);
-    let sk = secp256k1::SecretKey::from_slice(&data).unwrap();
-    let pk = secp256k1::PublicKey::from_secret_key(secp, &sk);
-    (sk, pk)
-}
-
 /// Generates a random P2PKH address.
 pub fn random_p2pkh_address(network: Network) -> Address {
-    with_rng(|rng| {
-        let secp = Secp256k1::new();
+    let secp = Secp256k1::new();
 
-        Address::p2pkh(&PublicKey::new(generate_keypair(&secp, rng).1), network)
-    })
+    Address::p2pkh(&PublicKey::new(generate_keypair(&secp).1), network)
 }
 
 pub fn random_p2tr_address(network: Network) -> Address {
-    with_rng(|rng| {
-        let secp = Secp256k1::new();
-        let (sk, _) = generate_keypair(&secp, rng);
-        let key_pair = KeyPair::from_secret_key(&secp, sk);
-        let xonly = XOnlyPublicKey::from_keypair(&key_pair);
+    let secp = Secp256k1::new();
+    let (sk, _) = generate_keypair(&secp);
+    let key_pair = KeyPair::from_secret_key(&secp, sk);
+    let xonly = XOnlyPublicKey::from_keypair(&key_pair);
 
-        Address::p2tr(&secp, xonly, None, network)
-    })
+    Address::p2tr(&secp, xonly, None, network)
 }
 
 pub fn random_p2wpkh_address(network: Network) -> Address {
-    with_rng(|rng| {
-        let secp = Secp256k1::new();
-        Address::p2wpkh(&PublicKey::new(generate_keypair(&secp, rng).1), network)
-            .expect("failed to create p2wpkh address")
-    })
+    let secp = Secp256k1::new();
+    Address::p2wpkh(&PublicKey::new(generate_keypair(&secp).1), network)
+        .expect("failed to create p2wpkh address")
 }
 
 pub fn random_p2wsh_address(network: Network) -> Address {
-    with_rng(|rng| {
-        let mut bytes = [0u8; 32];
-        rng.fill_bytes(&mut bytes);
-        Address::p2wsh(
-            &Script::new_v0_p2wsh(&WScriptHash::from_hash(Hash::from_slice(&bytes).unwrap())),
-            network,
-        )
-    })
+    let mut bytes = [0u8; 32];
+    fill_bytes(&mut bytes);
+    Address::p2wsh(
+        &Script::new_v0_p2wsh(&WScriptHash::from_hash(Hash::from_slice(&bytes).unwrap())),
+        network,
+    )
 }
 
 fn coinbase_input() -> TxIn {
